% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bubble.R
\name{geom_bubble}
\alias{geom_bubble}
\title{geom_bubble}
\usage{
geom_bubble(fill = "grey35", ...)
}
\arguments{
\item{fill}{character string color name or hexidecimal entry}

\item{...}{}
}
\value{
a ggplot2 layer
}
\description{
Like geom point but has a rim and a filling
}
\examples{
# without function
library(ggplot2)
ggplot(cars) +
  aes(x = speed, y = dist) +
  geom_point(shape = 21,size = 7,
    alpha = .85,
    color = "white",
    fill = "plum2")

 geom_bubble

# using function
library(ggplot2)
ggplot(cars) +
  aes(x = speed, y = dist) +
  geom_bubble(size = 7,
    alpha = .85,
    color = "white",
    fill = "plum2"
   ) +
 theme_minimal()


}
