% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_line_endpoint.R
\name{geom_line_endpoint}
\alias{geom_line_endpoint}
\title{geom_line_endpoint}
\usage{
geom_line_endpoint(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add a charming endpoint to your lines
}
\examples{
# functions
geom_line_endpoint
library(ggplot2)
ggplot(cars) +
   aes(x = speed, y = dist) +
   geom_line() +
   geom_line_endpoint()

ggplot(cars) +
    aes(x = speed, y = dist, color = dist > 15) +
    geom_line() +
    geom_line_endpoint()

ggplot(cars) +
     aes(x = speed, y = dist,
         color = dist > 15,
         linetype = speed > 16) +
     geom_line() +
     geom_line_endpoint()
}
