% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ring.R
\name{geom_ring}
\alias{geom_ring}
\title{geom_ring}
\usage{
geom_ring(...)
}
\arguments{
\item{...}{}

\item{fill}{character string color name or hexidecimal entry}
}
\value{
a ggplot2 layer
}
\description{
Like geom point but only has a rim by default
}
\examples{
# without function
library(ggplot2)
ggplot(cars) +
  aes(x = speed, y = dist) +
  geom_ring(size = 7, color = "slateblue")

 geom_ring

# using function
library(ggplot2)
ggplot(cars) +
  aes(x = speed, y = dist) +
  geom_ring(size = 7, color = "slateblue")
}
