% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_count.R
\name{geom_text_count}
\alias{geom_text_count}
\title{geom_text_count}
\usage{
geom_text_count(
  nudge_y = 0,
  position = ggplot2::position_dodge2(width = 0.9, preserve = "single"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add label which is the count.  Acts like geom_bar() but text label at the position of the top of the bar.
}
\examples{
library(ggplot2)
ggplot(mtcars) +
   aes(x = cyl) +
   geom_bar() +
   geom_text_count(nudge_y = .5)
}
