% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_count_percentage.R
\name{geom_text_count_percent}
\alias{geom_text_count_percent}
\title{geom_text_count_percent}
\usage{
geom_text_count_percent(
  nudge_y = 0,
  lineheight = 0.85,
  position = ggplot2::position_dodge2(width = 0.9, preserve = "single"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add label which is the count as well as percentage.  Acts like geom_bar() but text label at the position of the top of the bar.  Percentage is calculated within facet panels.
}
\examples{
library(ggplot2)
ggplot(mtcars) +
    aes(x = cyl) +
    geom_bar() +
    geom_text_count_percent(nudge_y = .5)
}
