% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_prop_over_time.R
\name{geom_tile_prop_over_time}
\alias{geom_tile_prop_over_time}
\title{Title}
\usage{
geom_tile_prop_over_time(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
library(magrittr)
gapminder::gapminder \%>\%
ggplot() +
  aes(x = year) +
  aes(y = continent) +
  geom_tile_prop_over_time() +
  aes(fill = pop) +
  scale_fill_viridis_c()
}
