% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_prop_over_time.R
\name{geom_tile_prop_over_time_text}
\alias{geom_tile_prop_over_time_text}
\title{Title}
\usage{
geom_tile_prop_over_time_text(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
library(magrittr)
gapminder::gapminder \%>\%
ggplot() +
  aes(x = year) +
  aes(y = continent) +
  geom_tile_prop_over_time() +
  aes(fill = pop) +
  scale_fill_viridis_c() +
  aes(label = pop) +
  geom_tile_prop_over_time_text()

gapminder::gapminder \%>\%
ggplot() +
  aes(x = year) +
  aes(y = continent) +
  geom_tile_prop_over_time() +
  aes(fill = pop) +
  scale_fill_viridis_c() +
  aes(label = pop) +
  geom_tile_prop_over_time_text() +
  facet_wrap(facets = vars(gdpPercap > 10000))

gapminder::gapminder \%>\%
ggplot() +
  aes(x = year) +
  aes(y = continent) +
  aes(fill = 1, label = 1) + # use 1 to count times continent is observed
  geom_tile_prop_over_time(color = "oldlace") +
  labs(fill = "proportion\nof countries\nin each time\nperiod") +
  geom_tile_prop_over_time_text()

}
